+++ lib/queued-graphql/servers/queued-subscription/queued-subscription-message.d.ts
@@ -1,0 +1,7 @@
+import { SubscriptionResponse } from './queued-subscription-request.interface';
+export interface QueuedSubscriptionMessage {
+    owner_id: string;
+    subscription_name: string;
+    subscription_id: string;
+    subscription_response: SubscriptionResponse;
+}