+++ src/queued-subscription-server/queued_subscription_request/db/index.ts
@@ -1,3 +1,0 @@
-/* eslint-disable import/no-cycle */
-export * from './queued-subscription-request.model';
-export * from './queued-subscription-request.service';